logger: Refactor into a separate module (iss #17)
